requirements.txt
augly==1.0.0
datasets==2.15.0
diffusers==0.24.0
easyocr==1.7.1
librosa==0.10.1
numpy==1.23.5
openai==0.27.9
opencv_python_headless==4.9.0.80
pandas==2.0.1
Pillow==10.2.0
python_dateutil==2.8.2
python_Levenshtein==0.25.0
Requests==2.31.0
scikit_learn==1.3.1
textract==1.6.5
torch==2.1.2
transformers==4.34.0
